\name{coralRG}
\alias{coralRG}
\docType{data}
\title{Spotted microarray red and green foreground and background values}
\description{
  Unnormalised red and green values, and corresponding background values.
  Further information is in the data frame \code{coralTargets}
}
\usage{data(coralRG)}
\format{
  The format is:
  Formal class 'RGList' [package "limma"]
  Can be accessed as a list, with named elements
"R" ,"G" ,"Rb" ,"Gb" ,"targets", "source" ,"genes" and "printer"
}
\source{
  Lauretto Grasso and Eldon Ball, Molecular Genetics and Evolution Group,
  Research School of Biological Sciences, Australian National University.
}
\examples{
data(coralRG)
}
\keyword{datasets}
